oldoffice CPU format, implement FMT_BLOB. It seems to be much harder doing it for the GPU format due to internal mask stuff.  See #3497
